% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{accordion}
\alias{accordion}
\title{Accordion}
\usage{
accordion(..., .id = NULL, .flush = FALSE)
}
\arguments{
\item{...}{Any number of \code{\link[=accordionItem]{accordionItem()}}.}

\item{.id}{Id of accordion.}

\item{.flush}{remove default background color, some
borders, some rounded corners.}
}
\description{
Collapsible accordion made up of \code{\link[=accordionItem]{accordionItem()}}.
}
\examples{
library(shiny)

ui <- fluidPage(
  theme = bslib::bs_theme(version = 5L),
  accordion(
    accordionItem(
      .expanded = TRUE,
      "Something",
      p("Some content")
    ),
    accordionItem(
      "Something else",
      p("Some other content")
    )
  )
)

server <- \(input, output, session){

}

if(interactive())
 shinyApp(ui, server)

}
