% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{accordionItem}
\alias{accordionItem}
\title{AccordionItem}
\usage{
accordionItem(
  label,
  ...,
  .id = NULL,
  .expanded = FALSE,
  .label_tag = shiny::h2
)
}
\arguments{
\item{label}{Label of the item.}

\item{...}{Elements to place in the body of the item.}

\item{.id}{Id of item.}

\item{.expanded}{Whether the element is initially expanded.}

\item{.label_tag}{Tag to use for the label.}
}
\description{
Accordion item to place within \code{\link[=accordion]{accordion()}}.
}
