% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alert}
\alias{alert}
\title{Alert}
\usage{
alert(..., class = "")
}
\arguments{
\item{...}{Content of alert.}

\item{class}{Additional classes, e.g.: \code{alert-success}.}
}
\description{
Alert
}
\examples{
alert(
  alertHeading("Header!"),
  shiny::p("Some content")
)

}
