% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alertHeading}
\alias{alertHeading}
\title{Alert Heading}
\usage{
alertHeading(heading)
}
\arguments{
\item{heading}{A character string or an HTML tag.}
}
\description{
Alert heading to be placed in \code{\link[=alert]{alert()}}.
}
