% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alertServer}
\alias{alertServer}
\alias{alert_show}
\alias{alert_hide}
\title{Toggle Alert}
\usage{
alert_show(id, session = shiny::getDefaultReactiveDomain())

alert_hide(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{ID of \code{\link[=alert]{alert()}} to toggle.}

\item{session}{A valid shiny session.}
}
\description{
Show or hide an alert.
}
