% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholders.R
\name{asPlaceholder}
\alias{asPlaceholder}
\title{As Placeholder}
\usage{
asPlaceholder(el, glow = FALSE)
}
\arguments{
\item{el}{Element to turn into a placeholder.}

\item{glow}{Whether the elements should glow.}
}
\description{
Turn an element into a placeholder.
}
\examples{
shiny::p() |> 
 asPlaceholder()

shiny::actionButton("id", "") |> 
 asPlaceholder()

}
