% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge}
\alias{badge}
\title{Badge}
\usage{
badge(
  ...,
  color = c("primary", "secondary", "success", "danger", "warning", "info", "light",
    "dark"),
  pill = FALSE,
  class = ""
)
}
\arguments{
\item{...}{Content of the badge.}

\item{color}{Background color of the badge.}

\item{pill}{Whether to create a rounded badge.}

\item{class}{Additional classes to passs to the badge.}
}
\description{
Badge.
}
