% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{buttonGroup}
\alias{buttonGroup}
\title{Button Group}
\usage{
buttonGroup(..., class = "")
}
\arguments{
\item{...}{Any \code{\link[shiny:actionButton]{shiny::actionButton()}}.}

\item{class}{Any additional class.}
}
\description{
Group of buttons.
}
\examples{
library(shiny)

ui <- fluidPage(
  theme = bslib::bs_theme(version = 5L),
  buttonGroup(
    actionButton("first", "First", class = "btn-primary"),
    actionButton("second", "Second", class = "btn-warning"),
    actionButton("third", "Third", class = "btn-success")
  )
)

server <- \(input, output, session){

}

if(interactive())
 shinyApp(ui, server)

}
