% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card}
\alias{card}
\title{Card}
\usage{
card(...)
}
\arguments{
\item{...}{Content, generally \code{\link[=cardBody]{cardBody()}}.}
}
\description{
Card.
}
\examples{
card(
  cardHeader("Header"),
  cardBody(
    cardTitle("Title"),
    cardSubtitle("Subtitle"),
    cardText("The content of the card")
  )
)

}
