% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{cardBody}
\alias{cardBody}
\title{Card Body}
\usage{
cardBody(...)
}
\arguments{
\item{...}{Content, HTML tags, \code{\link[=cardText]{cardText()}}, \code{\link[=cardTitle]{cardTitle()}},
or \code{\link[=cardSubtitle]{cardSubtitle()}}.}
}
\description{
Card body to be placed within \code{\link[=card]{card()}}.
}
