% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{cardHeader}
\alias{cardHeader}
\title{Card Header}
\usage{
cardHeader(content)
}
\arguments{
\item{content}{Content of the title or subtitle, either
a character vector or an HTML header tag.}
}
\description{
Card header placed within \code{\link[=card]{card()}} before \code{\link[=cardBody]{cardBody()}}.
}
