% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{cardTitle}
\alias{cardTitle}
\alias{cardSubtitle}
\title{Card Title & Subtitle}
\usage{
cardTitle(content)

cardSubtitle(content)
}
\arguments{
\item{content}{Content of the title or subtitle, either
a character vector or an HTML header tag.}
}
\description{
Card title and subtitle to be placed within \code{\link[=cardBody]{cardBody()}}.
}
