% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel.R
\name{carousel}
\alias{carousel}
\title{Carousel}
\usage{
carousel(..., id = NULL, class = "", controls = FALSE, indicators = FALSE)
}
\arguments{
\item{...}{Slides of the carousel, see \code{\link[=carouselItem]{carouselItem()}}.}

\item{id}{ID of carousel.}

\item{class}{Additional class of the carousel.}

\item{controls}{Whether to show control arrows.}

\item{indicators}{Whether to show indicators.}
}
\description{
Carousel element.
}
\examples{
library(shiny)

ui <- fluidPage(
  theme = bslib::bs_theme(version = 5L),
  carousel(
    indicators = TRUE,
    controls = TRUE,
    carouselItem(
      h5("Hello"),
      class = "bg-primary rounded p-4"
    ),
    carouselItem(
      h5("World"),
      class = "bg-info rounded p-4"
    )
  )
)

server <- \(input, output, session){
}

if(interactive())
 shinyApp(ui, server)

}
