% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel.R
\name{carouselControls}
\alias{carouselControls}
\title{Carousel Controls}
\usage{
carouselControls(show, id)
}
\arguments{
\item{show}{Whether to show the control arrows.}

\item{id}{ID of parent \code{\link[=carousel]{carousel()}}.}
}
\description{
Carousel control arrows.
}
\keyword{internal}
