% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel.R
\name{carouselImg}
\alias{carouselImg}
\title{Carousel Image}
\usage{
carouselImg(..., class = "")
}
\arguments{
\item{...}{Passed to \verb{<img/>}.}

\item{class}{Additional classes.}
}
\description{
Carousel image to use within \code{\link[=carouselItem]{carouselItem()}}.
}
