% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel.R
\name{carouselIndicators}
\alias{carouselIndicators}
\title{Carousel Indicators}
\usage{
carouselIndicators(items, id, show)
}
\arguments{
\item{items}{list of \code{\link[=carouselItem]{carouselItem()}}.}

\item{id}{Parent if of \code{\link[=carousel]{carousel()}}.}

\item{show}{Whether to show the indicators.}
}
\description{
Carousel indicators
}
\keyword{internal}
