% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\alias{collapse.collapseTrigger}
\title{Collapse}
\usage{
collapse(trigger, content)

\method{collapse}{collapseTrigger}(trigger, content)
}
\arguments{
\item{trigger}{A \code{\link[=collapseButton]{collapseButton()}} or \code{\link[=collapseLink]{collapseLink()}}.}

\item{content}{The offcanvas content as returned by
\code{\link[=offcanvasContent]{offcanvasContent()}}.}
}
\description{
Create a collapsible.
}
\section{Methods (by class)}{
\itemize{
\item \code{collapseTrigger}: Collapse wrapper
}}

\examples{
library(shiny)

ui <- fluidPage(
  theme = bslib::bs_theme(version = 5L),
  collapse(
    collapseLink(
      "Reveal"
    ),
    collapseContent(
      "Some collapsed content"
    )
  )
)

server <- \(input, output, session){

}

if(interactive())
 shinyApp(ui, server)

}
