% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapseContent}
\alias{collapseContent}
\title{Collapse Content}
\usage{
collapseContent(..., .id = NULL)
}
\arguments{
\item{...}{Content of the collapsible.}

\item{.id}{Id of the element.}
}
\description{
Collapse content, to pass to \code{\link[=collapseButton]{collapseButton()}} or
\code{\link[=collapseLink]{collapseLink()}}.
}
