% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapseLink}
\alias{collapseLink}
\title{Collapse Link}
\usage{
collapseLink(..., id = NULL, class = "")
}
\arguments{
\item{...}{Passed to the link.}

\item{id}{Link id.}

\item{class}{Additional class to pass to the link.}
}
\description{
Collapse link.
}
