% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{colorPickerInput}
\alias{colorPickerInput}
\title{Color Picker Input}
\usage{
colorPickerInput(inputId, value = NULL, ...)
}
\arguments{
\item{inputId}{ID of input.}

\item{value}{Initial value of the input.}

\item{...}{Passed to input.}
}
\description{
Color picker input.
}
