% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown.R
\name{dropdownButton}
\alias{dropdownButton}
\title{Dropdown Button}
\usage{
dropdownButton(label, ..., id = NULL, class = "")
}
\arguments{
\item{label}{Label of the dropdown.}

\item{...}{\code{\link[shiny:actionButton]{shiny::actionLink()}} to place within the
dropdown.}

\item{id}{ID of parent dropdown.}

\item{class}{Additional class for dropdown button.}
}
\description{
Dropdown buttons for \code{\link[shiny:actionButton]{shiny::actionLink()}}.
}
\examples{
library(shiny)

ui <- fluidPage(
  theme = bslib::bs_theme(version = 5L),
  dropdownButton(
    "Buttons",
    actionLink("first", "First"),
    actionLink("second", "Second")
  )
)

server <- \(input, output, session){

}

if(interactive())
  shinyApp(ui, server)

}
