% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{floatingSelectInput}
\alias{floatingSelectInput}
\title{Select Floating Label}
\usage{
floatingSelectInput(inputId, label, choices = NULL, ...)
}
\arguments{
\item{inputId}{ID of input.}

\item{label}{Label of input.}

\item{choices}{Choices, named vector.}

\item{...}{Passed to input.}
}
\description{
Select input with floating label.
}
