% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{floatingTextInput}
\alias{floatingTextInput}
\title{Text Input Floating Label}
\usage{
floatingTextInput(inputId, label, placeholder, value = "")
}
\arguments{
\item{inputId}{ID of input.}

\item{label}{Label of input.}

\item{placeholder}{Placeholder value.}

\item{value}{Value of input.}
}
\description{
Text input with floating label.
}
