% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{listGroup}
\alias{listGroup}
\title{List Group}
\usage{
listGroup(...)
}
\arguments{
\item{...}{Content of the list group, see \code{\link[=listGroupItem]{listGroupItem()}}.}
}
\description{
List group, see \code{\link[=listGroupItem]{listGroupItem()}}.
}
