% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{listGroupItem}
\alias{listGroupItem}
\title{List Group Item}
\usage{
listGroupItem(..., class = "")
}
\arguments{
\item{...}{Content.}

\item{class}{Additional classes.}
}
\description{
Items for \code{\link[=listGroup]{listGroup()}}.
}
