% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal}
\alias{modal}
\title{Modal}
\usage{
modal(
  id,
  ...,
  static_backdrop = FALSE,
  scrollable = FALSE,
  centered = FALSE,
  size = c("default", "sm", "lg", "xl", "fullscreen", "fullscreen-sm-down",
    "fullscreen-md-down", "fullscreen-lg-down", "fullscreen-xl-down",
    "fullscreen-xxl-down")
)
}
\arguments{
\item{id}{ID of modal.}

\item{...}{Content of the modal, \code{\link[=modalBody]{modalBody()}},
\code{\link[=modalHeader]{modalHeader()}}, \code{\link[=modalFooter]{modalFooter()}}.}

\item{static_backdrop}{When backdrop is set to static,
the modal will not close when clicking outside it.}

\item{scrollable}{Whether to make the content scrollable.}

\item{centered}{Whether to make the content vertically
centered.}

\item{size}{Size of the modal.}
}
\description{
Create a modal in the UI.
The modal is not down see \code{\link[=modal_show]{modal_show()}}, or
\code{\link[=modal_toggle]{modal_toggle()}} (from server).
}
\examples{
library(shiny)

ui <- fluidPage(
  theme = bslib::bs_theme(version = 5L),
  actionButton("show", "Show modal"),
  modal(
    id = "modal",
    modalHeader(
      modalTitle("Title")
    ),
    modalBody(
      p("The content of the modal")
    ),
    modalFooter(
      p("The footer of the modal")
    )
  )
)

server <- \(input, output, session){
  observeEvent(input$show, {
    modal_show("modal")
  })
}

if(interactive())
 shinyApp(ui, server)

}
