% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalFooter}
\alias{modalFooter}
\title{Modal Footer}
\usage{
modalFooter(...)
}
\arguments{
\item{...}{Content of the footer.}
}
\description{
Footer of the modal, place within \code{\link[=modal]{modal()}}.
}
