% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalHeader}
\alias{modalHeader}
\title{Modal Header}
\usage{
modalHeader(..., .close = TRUE)
}
\arguments{
\item{...}{Content.}

\item{.close}{Whether to add the close button.}
}
\description{
Header of a modal, place within \code{\link[=modal]{modal()}}.
}
