% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalTitle}
\alias{modalTitle}
\title{Modal Title}
\usage{
modalTitle(title)
}
\arguments{
\item{title}{An HTML tag or a character vector of length 1.}
}
\description{
Modal title, place within \code{\link[=modalHeader]{modalHeader()}}.
}
