% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalToggle}
\alias{modalToggle}
\alias{modal_show}
\alias{modal_hide}
\alias{modal_toggle}
\title{Modal Toggle}
\usage{
modal_show(id, session = shiny::getDefaultReactiveDomain())

modal_hide(id, session = shiny::getDefaultReactiveDomain())

modal_toggle(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{ID of the \code{\link[=modal]{modal()}}.}

\item{session}{A valid shiny session.}
}
\description{
Show, hide, or toggle a modal from the server.
}
