% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalTrigger}
\alias{modalTrigger}
\alias{modalButton}
\alias{modalLink}
\title{Modal Triggers}
\usage{
modalButton(target, ..., class = "")

modalLink(target, ...)
}
\arguments{
\item{target}{Target ID of the \code{\link[=modal]{modal()}} to show.}

\item{...}{Content of link or button, generally a
character vector of length 1.}

\item{class}{Additional class to pass to the button.}
}
\description{
Trigger open a modal with a button or a link,
avoids goind through the server.
}
