% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offcanvas.R
\name{offcanvasButton}
\alias{offcanvasButton}
\alias{offcanvasButtonWithTooltip}
\title{Offcanvas Button}
\usage{
offcanvasButton(..., id = NULL, class = "")

offcanvasButtonWithTooltip(
  ...,
  title,
  placement = c("top", "right", "bottom", "left"),
  id = NULL,
  class = ""
)
}
\arguments{
\item{...}{Passed to the button.}

\item{id}{ID of the button.}

\item{class}{Additional class to pass to the button.}

\item{title}{HTML tag or character vector of length 1.}

\item{placement}{Placement of the tooltip.}
}
\description{
Offcanvas button trigger.
}
