% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offcanvas.R
\name{offcanvasHeader}
\alias{offcanvasHeader}
\title{Canvas Header}
\usage{
offcanvasHeader(header)
}
\arguments{
\item{header}{An HTML tag or a character string.}
}
\description{
Header for \code{\link[=offcanvasContent]{offcanvasContent()}}.
}
