% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offcanvas.R
\name{offcanvasLink}
\alias{offcanvasLink}
\title{Offcanvas Link}
\usage{
offcanvasLink(..., id = NULL, class = "")
}
\arguments{
\item{...}{Passed to the button.}

\item{id}{ID of the link.}

\item{class}{Additional class to pass to the button.}
}
\description{
Offcanvas link trigger.
}
