% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offcanvas.R
\name{offcanvasServer}
\alias{offcanvasServer}
\alias{offcanvas_show}
\alias{offcanvas_hide}
\alias{offcanvas_toggle}
\title{Offcanvas Server}
\usage{
offcanvas_show(id, session = shiny::getDefaultReactiveDomain())

offcanvas_hide(id, session = shiny::getDefaultReactiveDomain())

offcanvas_toggle(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{ID of \code{\link[=offcanvasContent]{offcanvasContent()}}.}

\item{session}{Valid shiny session.}
}
\description{
Toggle the offcanvas from the server.
}
