% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progressBar}
\alias{progressBar}
\title{Progress Bar}
\usage{
progressBar(value = 0L, min = 0L, max = 100L, id = NULL, class = "")
}
\arguments{
\item{value}{Value of progress bar.}

\item{min, max}{Minimum and maximum values.}

\item{id}{ID of progress bar.}

\item{class}{Additional classes.}
}
\description{
A progress bar to pass to \code{\link[=progress]{progress()}}.
}
