% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{rangeInput}
\alias{rangeInput}
\title{Range Input}
\usage{
rangeInput(id, label, value = 0L, ..., min = 0L, max = 100L, step = 1L)
}
\arguments{
\item{id}{ID of input.}

\item{label}{Label of input.}

\item{value}{Initial value of input.}

\item{...}{Passed to input.}

\item{min, max}{Minimum and maximum values of the range.}

\item{step}{Step of range change.}
}
\description{
A range input similar to a slider.
}
