% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{switchInput}
\alias{switchInput}
\title{Switch Input}
\usage{
switchInput(inputId, label, checked = FALSE, ...)
}
\arguments{
\item{inputId}{ID of input.}

\item{label}{Label of input.}

\item{checked}{Whether the input is initialised checked.}

\item{...}{Passed to input.}
}
\description{
Switch input, similar to checkbox.
}
