% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{textGroupInput}
\alias{textGroupInput}
\title{Text Group Input}
\usage{
textGroupInput(
  inputId,
  label,
  value = "",
  placeholder = NULL,
  .position = c("left", "right")
)
}
\arguments{
\item{inputId}{ID of input.}

\item{label}{Label of input.}

\item{value}{Initial value of input.}

\item{placeholder}{Placeholder for input.}

\item{.position}{Position of the label.}
}
\description{
Text group input.
}
