% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{toast}
\alias{toast}
\title{Toast}
\usage{
toast(
  id,
  ...,
  class = "",
  position = c("top-right", "top-left", "bottom-right", "bottom-left")
)
}
\arguments{
\item{id}{ID of toast, used to show/hide from server.}

\item{...}{Content, generally \code{\link[=toastHeader]{toastHeader()}}, \code{\link[=toastBody]{toastBody()}}.}

\item{class}{Any additional classes.}

\item{position}{Position of the toast.}
}
\description{
Toast notifications.
}
\examples{
library(shiny)

ui <- fluidPage(
  theme = bslib::bs_theme(version = 5L),
  toast(
    "toast",
    toastHeader(
      strong(
        "Toast",
        class = "me-auto"
      )
    ),
    toastBody("The body of the toast!")
  ),
  h1("Hello"),
  actionButton("show", "Show toast"),
  actionButton("hide", "Hide toast")
)

server <- \(input, output, session){
  observeEvent(input$show, {
    toast_show("toast", auto_hide = FALSE)
  })

  observeEvent(input$hide, {
    toast_hide("toast")
  })
}

if(interactive())
 shinyApp(ui, server)

}
