% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{toastContent}
\alias{toastContent}
\alias{toastHeader}
\alias{toastBody}
\title{Toast Content}
\usage{
toastHeader(..., class = "")

toastBody(..., class = "")
}
\arguments{
\item{...}{Content.}

\item{class}{Any additional class.}
}
\description{
Content of the toast to place within \code{\link[=toast]{toast()}}.
}
