% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{toastToggle}
\alias{toastToggle}
\alias{toast_show}
\alias{toast_hide}
\title{Toast Toggle}
\usage{
toast_show(
  id,
  header = NULL,
  body = NULL,
  animation = TRUE,
  auto_hide = TRUE,
  delay = 5000L,
  session = shiny::getDefaultReactiveDomain()
)

toast_hide(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{ID of \code{\link[=toast]{toast()}} to toggle.}

\item{header, body}{Header and body of the toast
to replace existing content.}

\item{animation}{Whether to animate the toast.}

\item{auto_hide}{Whether to automatically hide the
toast after \code{delay}.}

\item{delay}{Milliseconds after which to hide the
toast (if \code{auto_hide = TRUE}).}

\item{session}{A valid shiny session.}
}
\description{
Show or hide a toast.
}
