% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{update_color_picker}
\alias{update_color_picker}
\title{Update Color Picker}
\usage{
update_color_picker(id, value, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{ID of input to update.}

\item{value}{Value to set.}

\item{session}{A valid shiny session.}
}
\description{
Update a \code{\link[=colorPickerInput]{colorPickerInput()}} from the server.
}
