% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{update_range}
\alias{update_range}
\title{Update Range}
\usage{
update_range(id, ..., session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{ID of input to update.}

\item{...}{Elements to update, one or more of,
\code{value}, \code{min}, and \code{max}.}

\item{session}{A valid shiny session.}
}
\description{
Update a \code{\link[=rangeInput]{rangeInput()}} from the server.
}
