% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{update_text_group}
\alias{update_text_group}
\title{Update Text Group Input}
\usage{
update_text_group(id, ..., session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{ID of input to update.}

\item{...}{Items to update; \code{value}, and/or \code{label}.}

\item{session}{Valid shiny session.}
}
\description{
Update \code{\link[=textGroupInput]{textGroupInput()}}.
}
