% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popover.R
\name{withPopover}
\alias{withPopover}
\title{Popover}
\usage{
withPopover(
  el,
  content,
  title = NULL,
  container = "body",
  placement = c("top", "right", "bottom", "left")
)
}
\arguments{
\item{el}{Element to place popover on.}

\item{content}{Content of the popover.}

\item{title}{Optional popover title.}

\item{container}{container of the popover.}

\item{placement}{Position of the popover.}
}
\description{
Add a popover to an element.
}
