% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{withTooltip}
\alias{withTooltip}
\title{Tooltip}
\usage{
withTooltip(el, title, placement = c("top", "right", "bottom", "left"))
}
\arguments{
\item{el}{Element.}

\item{title}{HTML tag or character vector of length 1.}

\item{placement}{Placement of the tooltip.}
}
\description{
Add a tooltip to an element.
}
\examples{
library(shiny)

ui <- fluidPage(
  theme = bslib::bs_theme(version = 5L),
  withTooltip(
    span("plain"),
    "content"
  ),
  br(),
  withTooltip(
    span("HTML"),
    tagList(
      strong("HTML"),
      "content"
    )
  )
)

server <- \(input, output, session){
}

if(interactive())
 shinyApp(ui, server)

}
