% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alert}
\alias{alert}
\title{Alert}
\usage{
alert(..., id = NULL, class = "")
}
\arguments{
\item{...}{Content of alert.}

\item{id}{ID of alert useful to toggle from server with
\code{\link[=alert_show]{alert_show()}} or \code{\link[=alert_hide]{alert_hide()}}.}

\item{class}{Additional classes, e.g.: \code{alert-success}.}
}
\description{
Alert
}
\examples{
alert(
  alertHeading("Header!"),
  shiny::p("Some content")
)

}
