% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel.R
\name{carouselItem}
\alias{carouselItem}
\title{Carousel Item}
\usage{
carouselItem(..., interval = NULL, class = "")
}
\arguments{
\item{...}{Content of item.}

\item{interval}{Number of milliseconds this item should be shown.}

\item{class}{Additional classes.}
}
\description{
Carousel item to use within \code{\link[=carousel]{carousel()}}.
}
