% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapseButton}
\alias{collapseButton}
\title{Collapse Button}
\usage{
collapseButton(..., id = NULL, class = "")
}
\arguments{
\item{...}{Passed to the button.}

\item{id}{Button id.}

\item{class}{Additional classes to pass to the button.}
}
\description{
Collapse button.
}
\examples{
library(shiny)

ui <- fluidPage(
  theme = bslib::bs_theme(version = 5L),
  collapse(
    collapseButton(
      "Reveal"
    ),
    collapseContent(
      "Some collapsed content"
    )
  )
)

server <- \(input, output, session){

}

if(interactive())
 shinyApp(ui, server)

}
