% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapseButton}
\alias{collapseButton}
\alias{collapseButton.collapseContent}
\title{Collapse Button}
\usage{
collapseButton(content, ..., class = "")

\method{collapseButton}{collapseContent}(content, ..., class = "")
}
\arguments{
\item{content}{An object of class \code{collapseContent}
as returned by \code{\link[=collapseContent]{collapseContent()}}.}

\item{...}{Passed to the button.}

\item{class}{Additional classes to pass to the button.}
}
\description{
Collapse button.
}
\section{Methods (by class)}{
\itemize{
\item \code{collapseContent}: Collapse button
}}

\examples{
library(shiny)

ui <- fluidPage(
  theme = bslib::bs_theme(version = 5L),
  collapseButton(
    collapseContent(
      "Some collapsed content"
    ),
    "Reveal"
  )
)

server <- \(input, output, session){

}

if(interactive())
 shinyApp(ui, server)

}
