% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapseLink}
\alias{collapseLink}
\alias{collapseLink.collapseContent}
\title{Collapse Link}
\usage{
collapseLink(content, ...)

\method{collapseLink}{collapseContent}(content, ...)
}
\arguments{
\item{content}{An object of class \code{collapseContent}
as returned by \code{\link[=collapseContent]{collapseContent()}}.}

\item{...}{Passed to the link.}
}
\description{
Collapse link.
}
\section{Methods (by class)}{
\itemize{
\item \code{collapseContent}: Collapse button
}}

