% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{floatingTextAreaInput}
\alias{floatingTextAreaInput}
\title{Text Input Area Floating Label}
\usage{
floatingTextAreaInput(inputId, label, placeholder, value = "", ...)
}
\arguments{
\item{inputId}{ID of input.}

\item{label}{Label of input.}

\item{placeholder}{Placeholder value.}

\item{value}{Value of input.}

\item{...}{Passed to input.}
}
\description{
Text input with floating label.
}
