% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offcanvas.R
\name{offcanvasButton}
\alias{offcanvasButton}
\alias{offcanvasButton.offcanvasContent}
\title{Offcanvas Button}
\usage{
offcanvasButton(content, ..., class = "")

\method{offcanvasButton}{offcanvasContent}(content, ..., class = "")
}
\arguments{
\item{content}{Content of the offcanvas as returned by
\code{\link[=offcanvasContent]{offcanvasContent()}}.}

\item{...}{Passed to the button.}

\item{class}{Additional class to pass to the button.}
}
\description{
Offcanvas button trigger.
}
\section{Methods (by class)}{
\itemize{
\item \code{offcanvasContent}: Button for offcanvas content
}}

\examples{
library(shiny)

ui <- fluidPage(
  theme = bslib::bs_theme(version = 5L),
  offcanvasButton(
    offcanvasContent(
      offcanvasHeader(
        "Off canvas"
      ),
      p(
        "Hello world"
      )
    ),
    "Open"
  )
)

server <- \(input, output, session){

}

if(interactive())
 shinyApp(ui, server)

}
