% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offcanvas.R
\name{offcanvasContent}
\alias{offcanvasContent}
\title{Offcanvas Content}
\usage{
offcanvasContent(
  ...,
  id = NULL,
  class = "",
  .position = c("start", "end", "top", "bottom"),
  scroll = TRUE
)
}
\arguments{
\item{...}{Content of the offcanvas.}

\item{id}{Id of the canvas.}

\item{class}{Additional classes.}

\item{.position}{Where the canvas should open.}

\item{scroll}{Whether to allow scrolling the body.}
}
\description{
Offcanvas content.
}
